+++ packages/dnd/src/isTouchDevice.ts
@@ -1,4 +1,4 @@
-// utils.ts
+/* istanbul ignore file */
 export const isTouchDevice = (): boolean =>
   (typeof window !== 'undefined' && 'ontouchstart' in window) ||
   (typeof navigator !== 'undefined' && navigator.maxTouchPoints > 0);