+++ packages/native/src/components/defaults.ts
@@ -3,6 +3,8 @@ import type { Controls, FullField } from 'react-querybuilder';
 import { RuleGroupBodyComponents, RuleGroupHeaderComponents } from 'react-querybuilder';
 import { NativeActionElement } from './NativeActionElement';
 import { NativeInlineCombinator } from './NativeInlineCombinator';
+import { NativeMatchModeEditor } from './NativeMatchModeEditor';
+import { NativeMatchModeEditorWeb } from './NativeMatchModeEditorWeb';
 import { NativeNotToggle } from './NativeNotToggle';
 import { NativeShiftActions } from './NativeShiftActions';
 import { NativeValueEditor } from './NativeValueEditor';
@@ -37,12 +39,14 @@ export const defaultNativeControlElements: Controls<FullField, string> = {
   valueEditor: NativeValueEditor,
   valueSelector: NativeValueSelector,
   valueSourceSelector: NativeValueSelector,
+  matchModeEditor: NativeMatchModeEditor,
 };
 
 export const defaultNativeWebControlElements: Controls<FullField, string> = {
   ...defaultNativeControlElements,
   combinatorSelector: NativeValueSelectorWeb,
   fieldSelector: NativeValueSelectorWeb,
+  matchModeEditor: NativeMatchModeEditorWeb,
   operatorSelector: NativeValueSelectorWeb,
   valueEditor: NativeValueEditorWeb,
   valueSelector: NativeValueSelectorWeb,