+++ packages/react-querybuilder/src/types/basic.ts
@@ -1,11 +1,12 @@
-import type { SetOptional, Simplify } from './type-fest';
 import type {
   BaseFullOption,
+  FlexibleOption,
   FlexibleOptionList,
   FullOption,
-  Option,
+  StringUnionToFullOptionArray,
   WithUnknownIndex,
 } from './options';
+import type { SetOptional, Simplify } from './type-fest';
 import type { RuleValidator } from './validation';
 
 /**
@@ -84,6 +85,22 @@ export type InputType =
   | 'week'
   | (string & {});
 
+/**
+ * Quantification mode describing how many elements of the value array must pass
+ * the filter for the rule itself to pass.
+ *
+ * For "atLeast", "atMost", and "exactly", the threshold value will be converted to
+ * a percentage if the number is less than 1. Non-numeric values and numbers less
+ * than 0 will be ignored.
+ */
+export interface MatchConfig {
+  mode: MatchMode;
+  threshold?: number | null | undefined;
+}
+
+export type MatchMode = 'all' | 'some' | 'none' | 'atLeast' | 'atMost' | 'exactly';
+export type MatchModeOptions = StringUnionToFullOptionArray<MatchMode>;
+
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type ActionElementEventHandler = (event?: any, context?: any) => void;
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -96,8 +113,8 @@ interface BaseFullField<
   FieldName extends string = string,
   OperatorName extends string = string,
   ValueName extends string = string,
-  OperatorObj extends Option = Option<OperatorName>,
-  ValueObj extends Option = Option<ValueName>,
+  OperatorObj extends FullOption = FullOption<OperatorName>,
+  ValueObj extends FullOption = FullOption<ValueName>,
 > extends WithOptionalClassName<BaseFullOption<FieldName>> {
   id?: string;
   operators?: FlexibleOptionList<OperatorObj>;
@@ -105,6 +122,9 @@ interface BaseFullField<
   valueSources?: ValueSources | ((operator: OperatorName) => ValueSources);
   inputType?: InputType | null;
   values?: FlexibleOptionList<ValueObj>;
+  matchModes?: boolean | MatchMode[] | FlexibleOption<MatchMode>[];
+  /** Properties of items in the value. */
+  subproperties?: FlexibleOptionList<FullField>;
   defaultOperator?: OperatorName;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   defaultValue?: any;
@@ -129,8 +149,8 @@ export type FullField<
   FieldName extends string = string,
   OperatorName extends string = string,
   ValueName extends string = string,
-  OperatorObj extends Option = Option<OperatorName>,
-  ValueObj extends Option = Option<ValueName>,
+  OperatorObj extends FullOption = FullOption<OperatorName>,
+  ValueObj extends FullOption = FullOption<ValueName>,
 > = Simplify<
   FullOption<FieldName> & BaseFullField<FieldName, OperatorName, ValueName, OperatorObj, ValueObj>
 >;
@@ -149,7 +169,7 @@ export type Field<
   FieldName extends string = string,
   OperatorName extends string = string,
   ValueName extends string = string,
-  OperatorObj extends Option = Option<OperatorName>,
+  OperatorObj extends FullOption = FullOption<OperatorName>,
 > = WithUnknownIndex<
   { value?: FieldName } & Pick<
     BaseFullField<FieldName, OperatorName, ValueName, OperatorObj>,
@@ -171,7 +191,7 @@ export type FieldByValue<
   FieldName extends string = string,
   OperatorName extends string = string,
   ValueName extends string = string,
-  OperatorObj extends Option = Option<OperatorName>,
+  OperatorObj extends FullOption = FullOption<OperatorName>,
 > = WithUnknownIndex<
   { name?: FieldName } & Pick<
     BaseFullField<FieldName, OperatorName, ValueName, OperatorObj>,