+++ packages/react-querybuilder/src/types/options.ts
@@ -1,5 +1,9 @@
 import type { RequireAtLeastOne, SetRequired, Simplify } from './type-fest';
 
+export type StringUnionToFullOptionArray<Op extends string> = Array<
+  Op extends unknown ? FullOption<Op> : never
+>;
+
 /**
  * Extracts the {@link Option} type from a {@link FlexibleOptionList}.
  *