+++ packages/react-querybuilder/src/utils/parseSQL/utils.ts
@@ -1,4 +1,5 @@
 import type { DefaultCombinatorNameExtended, DefaultOperatorName } from '../../types/index.noReact';
+import { lc } from '../misc';
 import { parseNumber } from '../parseNumber';
 import type { ParseSQLOptions } from './parseSQL';
 import type {
@@ -60,7 +61,7 @@ export const getFieldName = (f: string | SQLIdentifier): string => {
 };
 
 const normalizeCombinator = (c: AndOperator | OrOperator | XorOperator) =>
-  c.replace('&&', 'and').replace('||', 'or').toLowerCase() as DefaultCombinatorNameExtended;
+  lc(c.replace('&&', 'and').replace('||', 'or')) as DefaultCombinatorNameExtended;
 
 export const normalizeOperator = (op: ComparisonOperator, flip?: boolean): DefaultOperatorName => {
   if (flip) {
@@ -91,7 +92,7 @@ export const evalSQLLiteralValue = (
     // single or double quotes, but if we do get here we return the string as is.
     return valueString;
   } else if (valueObj.type === 'Boolean') {
-    return valueObj.value.toLowerCase() === 'true';
+    return lc(valueObj.value) === 'true';
   } else if (isSQLSignedNumber(valueObj)) {
     return parseNumber(`${valueObj.prefix}${valueObj.value.value}`, {
       parseNumbers: true,